# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Health & Nutrition Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meals:
    get:
      operationId: getMeals
      tags: [meals]
      summary: List meals for a date
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Log a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
...
  /meals/{id}:
    get:
      operationId: getMeal
      tags: [meals]
      summary: Get a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
    patch:
      operationId: updateMeal
      tags: [meals]
      summary: Update a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
    delete:
      operationId: deleteMeal
      tags: [meals]
      summary: Delete a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /food-items:
    post:
      operationId: createFoodItem
      tags: [food-items]
      summary: Add a food item to a meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
  /food-items/{id}:
    patch:
      operationId: updateFoodItem
      tags: [food-items]
      summary: Update a food item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
    delete:
      operationId: deleteFoodItem
      tags: [food-items]
      summary: Delete a food item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Meal:
      type: object
      required: [id, date, mealType]
      properties:
        id:
          type: number
        date:
          type: string
        mealType:
          type: string
        time:
          type: ["string", "null"]
        totalCalories:
          type: ["number", "null"]
        totalProtein:
          type: ["number", "null"]
        totalCarbs:
          type: ["number", "null"]
        totalFat:
          type: ["number", "null"]
        totalFiber:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        postMealSleepiness:
          type: ["number", "null"]
        foodItems:
          type: array
          items:
            $ref: "#/components/schemas/FoodItem"
        createdAt:
          type: string
    MealInput:
      type: object
      required: [date, mealType]
      properties:
        date:
          type: string
        mealType:
          type: string
        time:
          type: ["string", "null"]
        totalCalories:
          type: ["number", "null"]
        totalProtein:
          type: ["number", "null"]
        totalCarbs:
          type: ["number", "null"]
        totalFat:
          type: ["number", "null"]
        totalFiber:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        postMealSleepiness:
          type: ["number", "null"]
        foodItems:
          type: array
          items:
            $ref: "#/components/schemas/FoodItemInput"
    MealUpdate:
      type: object
      properties:
        mealType:
          type: ["string", "null"]
        time:
          type: ["string", "null"]
        totalCalories:
          type: ["number", "null"]
        totalProtein:
          type: ["number", "null"]
        totalCarbs:
          type: ["number", "null"]
        totalFat:
          type: ["number", "null"]
        totalFiber:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        postMealSleepiness:
          type: ["number", "null"]
...
    FoodItem:
      type: object
      required: [id, mealId, name]
      properties:
        id:
          type: number
        mealId:
          type: number
        name:
          type: string
        portion:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        calories:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
    FoodItemInput:
      type: object
      required: [mealId, name]
      properties:
        mealId:
          type: number
        name:
          type: string
        portion:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        calories:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
    FoodItemUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        portion:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        calories:
          type: ["number", "null"]
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        fiber:
          type: ["number", "null"]
...